components belonging to a specific card on single job page moved to a folder for their card. Parent component in each folder renamed to index.js
